refactor(suite): extract earn form hooks and move earn modals
