+++ packages/suite/src/components/earn/forms/ChangeDelegateFormContext.ts
@@ -1,0 +1,13 @@
+import type { FormState as ReactHookFormState, UseFormReturn } from 'react-hook-form';
+
+import {
+    BaseStakeContextValues,
+    ChangeDelegateFormState,
+    StakeFormState,
+} from '@suite-common/wallet-types';
+
+export type ChangeDelegateContextValues = UseFormReturn<ChangeDelegateFormState> &
+    BaseStakeContextValues & {
+        methods: UseFormReturn<ChangeDelegateFormState>;
+        formState: ReactHookFormState<StakeFormState>;
+    };
