+++ packages/suite/src/components/earn/modals/EarnSupplyModal/SupplyForm/CardanoStakeWarningBanner.tsx
