+++ packages/suite/src/components/earn/modals/EarnSupplyModal/SupplyForm/StakeRegistrationDepositCard.tsx
