+++ packages/suite/src/components/wallet/stakeForm/StakeForm.ts
@@ -1,13 +1,1 @@
-import type { FormState as ReactHookFormState, UseFormReturn } from 'react-hook-form';
-
-import {
-    BaseStakeContextValues,
-    ChangeDelegateFormState,
-    StakeFormState,
-} from '@suite-common/wallet-types';
-
-export type ChangeDelegateContextValues = UseFormReturn<ChangeDelegateFormState> &
-    BaseStakeContextValues & {
-        methods: UseFormReturn<ChangeDelegateFormState>;
-        formState: ReactHookFormState<StakeFormState>;
-    };
+export type { ChangeDelegateContextValues } from 'src/components/earn/forms/ChangeDelegateFormContext';
