+++ packages/suite/src/types/earn/earnFormFields.ts
@@ -1,0 +1,3 @@
+export const FIAT_INPUT = 'fiatInput';
+export const CRYPTO_INPUT = 'cryptoInput';
+export const OUTPUT_AMOUNT = 'outputs.0.amount';
