+++ packages/suite/src/types/wallet/claimForm.ts
@@ -1,16 +1,1 @@
-import { UseFormReturn } from 'react-hook-form';
-
-import { BaseStakeContextValues, StakeFormState } from '@suite-common/wallet-types';
-
-export type ClaimFormState = Omit<
-    StakeFormState,
-    'amountLimits' | 'onCryptoAmountChange' | 'onFiatAmountChange' | 'setMaxOutputId' | 'fiatInput'
->;
-
-export type ClaimContextValues = UseFormReturn<ClaimFormState> &
-    BaseStakeContextValues & {
-        onClaimChange: (amount: string) => Promise<void>;
-    } & {
-        methods: UseFormReturn<ClaimFormState>;
-        isClaimingDisabled?: boolean;
-    };
+export type { ClaimContextValues, ClaimFormState } from 'src/types/earn/claimForm';
