+++ packages/suite/src/types/wallet/stakeForms.ts
@@ -1,3 +1,1 @@
-export const FIAT_INPUT = 'fiatInput';
-export const CRYPTO_INPUT = 'cryptoInput';
-export const OUTPUT_AMOUNT = 'outputs.0.amount';
+export { CRYPTO_INPUT, FIAT_INPUT, OUTPUT_AMOUNT } from 'src/types/earn/earnFormFields';
